+++ src/main/scala/ru/sgu/switchmap/models/Floor.scala
@@ -2,7 +2,7 @@ package ru.sgu.switchmap.models
 
 final case class DBFloor(number: Int, buildName: String, buildShortName: String)
 
-final case class Floor(number: Int, switchesNumber: Int)
+final case class Floor(number: Int, switchesNumber: Long)
 
 final case class FloorNotFound(number: Int, buildShortName: String)
     extends Exception
